selftests: mptcp: tune timeout and delay for simult_flows cases
We currently have some instabilities in the simult_flows tests-case.
The problem boils down to the unneeded large wait introduced by the
tests to allow for the MPJ handshake to complete, which can also
introduce a quite relevant variance.

Do wait on a single side of the connection, remove the delay at
shutdown time and tune the expected test time with the above.

Closes: https://github.com/multipath-tcp/mptcp_net-next/issues/137
Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Message-Id: <126a63a07de5709e2702abb5f431f9128146c60a.1630922799.git.pabeni@redhat.com>